+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-16 11:51:07
-// Commit: docs(state): update readme and API contracts for state
+// Updated at: 2026-08-16 13:38:24
+// Commit: chore(state): update dependencies and cleanup unused imports in state
 
 class stateComponentState {
-  final String commitRef = '3636';
-  final String lastMessage = 'docs(state): update readme and API contracts for state';
-  final DateTime timestamp = DateTime.parse('2026-08-16T11:51:07+05:00');
+  final String commitRef = '3639';
+  final String lastMessage = 'chore(state): update dependencies and cleanup unused imports in state';
+  final DateTime timestamp = DateTime.parse('2026-08-16T13:38:24+05:00');
 }